Refactor userConfig to custom hook
