+++ README.md
@@ -1,6 +1,11 @@
 # KoieneRes Frontend
 
+## Status
+
+[![CircleCI](https://circleci.com/gh/NTNUIKoiene/koieneres-frontend/tree/master.svg?style=svg)](https://circleci.com/gh/NTNUIKoiene/koieneres-frontend/tree/master)
 [![Greenkeeper badge](https://badges.greenkeeper.io/NTNUIKoiene/koieneres-frontend.svg)](https://greenkeeper.io/)
+
+## About
 
 KoieneRes Frontend is the frontend application for the new reservation system for NTNUI Koiene. For the backend see the [here](https://github.com/NTNUIKoiene/koieneres-backend). An instance of the backend is running on [here](<[https://koieneres-api-dev.appspot.com/](https://koieneres-api-dev.appspot.com/api/)>). Swagger docs can be found [here](https://koieneres-api-dev.appspot.com/swagger/).
